pyproject.toml
[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

[project]
name = "chain-semantic-splitter"
version = "0.1.0"
description = "A LangChain TextSplitter that uses semantic understanding to intelligently merge text chunks."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Sitraka Herivalsoa", email = "" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: General",
]
dependencies = [
    "build>=1.3.0",
    "langchain-core>=0.1.40,<0.2.0",
    "langchain-google-genai>=0.0.9,<0.1.0",
    "twine>=6.0.1",
]

[project.urls]
Homepage = "https://github.com/sitrakaherivalisoagithub/chain-semantic-splitter.git"
Repository = "https://github.com/sitrakaherivalisoagithub/chain-semantic-splitter.git"

[tool.hatch.build.targets.wheel]
packages = ["src/chain_semantic_splitter"]
